4.21 valmis

+++ osa4/controllers/blogs.js
@@ -1,8 +1,17 @@
 const blogsRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
 const Blog = require('../models/blog');
 const User = require('../models/user');
 
-blogsRouter.get('/', async (req, res) => {
+/* const getTokenFrom = request => {
+  const authorization = request.get('authorization');
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7);
+  }
+  return null;
+}; */
+
+blogsRouter.get('/', async (req, res, next) => {
   try {
     const blogs = await Blog.find({}).populate('user', {
       username: 1,
@@ -10,37 +19,42 @@ blogsRouter.get('/', async (req, res) => {
     });
     res.json(blogs.map(blog => blog.toJSON()));
   } catch (err) {
-    console.error(err);
+    next(err);
   }
 });
 
-blogsRouter.post('/', async (req, res) => {
+blogsRouter.post('/', async (req, res, next) => {
   if (!req.body.title || !req.body.url) {
     res.status(400).end();
     return;
   }
 
-  const user = await User.findOne();
-
-  const blog = new Blog({
-    title: req.body.title,
-    author: req.body.author,
-    url: req.body.url,
-    likes: req.body.likes || 0,
-    user: user._id
-  });
-
   try {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!req.token || !decodedToken.id) {
+      return res.status(401).json({ error: 'Token missing or invalid' });
+    }
+
+    const user = await User.findById(decodedToken.id);
+
+    const blog = new Blog({
+      title: req.body.title,
+      author: req.body.author,
+      url: req.body.url,
+      likes: req.body.likes || 0,
+      user: user._id
+    });
+
     const savedBlog = await blog.save();
     user.blogs = user.blogs.concat(savedBlog._id);
     await user.save();
     res.status(201).json(savedBlog.toJSON());
   } catch (err) {
-    console.error(err);
+    next(err);
   }
 });
 
-blogsRouter.put('/:id', async (req, res) => {
+blogsRouter.put('/:id', async (req, res, next) => {
   if (!req.body.likes) {
     res.status(400).end();
     return;
@@ -56,16 +70,27 @@ blogsRouter.put('/:id', async (req, res) => {
     });
     res.json(updatedBlog.toJSON());
   } catch (err) {
-    console.error(err);
+    next(err);
   }
 });
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', async (req, res, next) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id);
-    res.status(204).end();
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!req.token || !decodedToken.id) {
+      return res.status(401).json({ error: 'Token missing or invalid' });
+    }
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(req.params.id);
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(req.params.id);
+      res.status(204).end();
+    } else {
+      res.status(401).end();
+    }
   } catch (err) {
-    console.error(err);
+    next(err);
   }
 });
